% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedaML.R
\name{visualise}
\alias{visualise}
\title{Complete Exploratory Data Analysis and Machine Learning Tool}
\usage{
visualise()
}
\arguments{
\item{NULL}{}
}
\description{
This package allows you to do EDA and run supervised machine Learning models via a UI in Shiny
}
\examples{
visualise()

}
